requirements.txt
SpeechRecognition
rich
pydub
requests
python-dotenv